setup_help.py into setup.py
